adapt CsvimIT.java
