+++ tests/tests-integrations/src/test/resources/CsvimIT/csvim/readers.csv
