+++ tests/tests-integrations/src/test/resources/CsvimIT/project.json
@@ -1,5 +1,5 @@
 {
-  "guid": "csvim-test-project",
+  "guid": "CsvimIT",
   "dependencies": [],
   "actions": [
     {
