+++ tests/tests-integrations/src/test/resources/CsvimIT/tables/reader.table
